requirements.txt
asgiref==3.4.1
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.9
django-cors-headers==3.10.0
django-filter==21.1
django-oauth-toolkit==1.5.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
drf-social-oauth2==1.0.9
drf-spectacular==0.20.2
drf-spectacular-sidecar==2021.11.2
idna==3.3
inflection==0.5.1
jsonschema==4.1.2
jwcrypto==1.0
oauthlib==3.1.1
Pillow==8.4.0
pycparser==2.20
PyJWT==2.3.0
pyrsistent==0.18.0
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
wrapt==1.13.3
mysqlclient==2.0.3